Fix platform cleanup for CaaS clusters (#185)
Identity platforms are not explicitly linked to CaaS clusters via
ownership, like they are for the Kubernetes side. This can lead to
situations where a cluster is deleted but the identity platform is left
behind. This causes alerts about the missing Zenith services until the
platform is manually removed. This is especially an issue when using the
auto-delete functionality, but can also cause issues when the identity
platform is not cleanly removed.

The solution is to create CaaS clusters in the same namespace as the
identity platform resources so that proper owner references can be used
to clean up the associated platform when a CaaS cluster is deleted.
Protections are put in place to make sure that namespaces cannot be
stomped on when projects change name etc.