Error on multiple networks with the same tag
This helps us be more consistent with CAPO that will error when it looks
for portal-internal and sees multiple networks.

Possibly more importantly, it stops surpizes when you accidentally end
up with mutiple networks with the same tag, and we end up relying on the
order of the networks being returned from the Neutron API's default sort
ordering.