Add ssh_key_is_public keystore property
ssh_key_is_public determines whether the UI should
make the SSH public key visible in the Update SSH key
modal.

It makes sense to display the current SSH key for both the
Provider and LDAP keystores, but as the SSH key for the
dummy keystore is provided by the Azimuth operator, its
value is not relevant to users (and may be confusing), so
there is no need to display it in the UI.

The default behaviour for all keystores is unchanged, but
DummyKeyStore may be configured to not show its SSH key
in the Azimuth UI by setting ssh_key_is_public: False.

Additionally, fix some logic around validateSSHKey in the
UI, which was incorrectly firing when an keystore that does
not support key update was in use.