Allow portal-internal to be a shared network
Add a new config option to the openstack provider -
allowSharedInternalNet - which allows a network shared into a
project to be used as the portal-internal network.

If allowSharedInternalNet is True, shared networks which are
available to the project and have the portal-internal tag are
preferentially chosen as the portal-internal network over those
that are owned by the project.

If allowSharedInternalNet is True and there is exactly one
shared network with the portal-internal tag available to the
project, then this network is chosed as the internal network.
If there is no shared network, but exactly one network owned
by the project with a portal-internal tag, this network is
chosen as the internal network. Multiple shared or project
networks with the portal-internal tag will result in an
error.

There are no changes to the default behaviour, and projects
that do not have access to a shared network with the
portal-internal tag will use an existing network owned by the
project if it is available if allowSharedInternalNet is False
(the default), or a new network will be created if
createInternalNet is True (the default).

Further, if no networks with the portal-internal tag are found,
the same rules are applied to network detection using
internalNetTemplate - shared networks are considered first, then
networks owned by the project.