Allow automatic tenant network creation to return None
If machine support is disabled for a provider, allow the
automated tenant network detection to return None and
not raise an error if a tenant network can't be found by
tag or name, and creation is disabled.

This implies that the cloud provider (and potentially the
tenant in the future) supports only Platforms, and that
connecting instances created as part of Platforms to the
appropriate networks is the responsibility of the
Platform, and not of Azimuth.