Dynamically allow tenancies access to Machines, Volumes, Kubernetes and Kubernetes Apps based on discovery of the portal-internal network  (#425)
### Internal network is a shared network
Add a new config option to the openstack provider -
`internalNetIsShared` - which allows a network shared into a project to
be used as the portal-internal network.

If `internalNetIsShared` is True, shared networks that are available to
the project and have the `portal-internal` tag are preferentially chosen
as the portal-internal network over those that are owned by the project.

There are no changes to the default behaviour, and projects that do not
have access to a **shared** network with the `portal-internal` tag will
use an existing network owned by the project if it is available, or a
new network will be created if `createInternalNet` is True (the
default).

### Internal network doesn't exist
If the internal network can not be detected and `createInternalNet` is
False, the capabilites returned for a scoped session are updated to
disable machines, volumes, Kubernetes and Kubernetes apps API endpoints,
because creating resources of these types relies on the internal network
being determined. The UI reacts to the disabling of these endpoints by
removing the relevant items from the sidebar (Advanced>Machines,
Volumes) or from the platforms catalogue (Kubernetes and Kubernetes
apps).

### Machines functionality disabled
The "machines" functionality may be disabled for the OpenStack provider
by setting `supportsMachines` to False (default is True). This will
disable operations on machines and volumes across all projects, and will
cause the "advanced" menu element to be removed from the UI.

### Dynamic project capabilities vs configured capabilities
Configuration options remain for enabling and disabling CaaS, Kubernetes
and Kubernetes apps, alongside a new configuration option for Machines.
If a capability is disabled in config, it will not be enabled
dynamically on a per-tenant basis. If a capability is enabled in config,
it may be disabled on a per-tenant basis if the prerequisites for the
capability are not present (for example, the internal network was not
able to be determined for a project, so machines, volumes, Kubernetes
and Kubernetes Apps are dynamically disabled even though they are
enabled in configuration).